General outline of the project
I created a general structure for this project and README file is created as well. I uploaded a file inside code folder for fetching data. You guys can ignore the write command if you don't want to make it local. Also the three raw datasets are inside the data folder.