qmd file is ready and render preview is set
I've finished majority of the stuff we need for qmd file and rendered for result. The only thing missing is the chart coding part. The reference part is also done and references are included at the end of the qmd file, after rendering, they will pop up under the references section in the pdf.