Fixed data_visualizations.R and finalized pdf
The visualizations.R file has many errors, the plots are not clear enough. The axis label are squeezed, height, width ratio are not correct and readability is poor. I added color blind friendly package for the plot as well. The figures folder is now up to date with the new figures.

final_project.pmd file is finalized, all labels for tables and figures are added, narrative text refers the correct ones.

Please read through the final_project.pdf file, README.md and check if any error or missing analysis exists.